[CLEANUP] Rename the Tests namespace to Testing
Closes #480
