+++ ext_localconf.php
@@ -2,4 +2,4 @@
 defined('TYPO3_MODE') or die('Access denied.');
 
 $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['oelib']['testingFrameworkCleanUp'][]
-    = \OliverKlee\Oelib\Tests\TestingFrameworkCleanup::class;
+    = \OliverKlee\Oelib\Testing\TestingFrameworkCleanup::class;
